Implements an abstract BasePathEditorControl
