+++ Xamarin.PropertyEditing.Mac/Xamarin.PropertyEditing.Mac.csproj
@@ -161,6 +161,7 @@
     <Compile Include="PropertyList.cs" />
     <Compile Include="Controls\FilePathEditorControl.cs" />
     <Compile Include="Controls\Custom\TextFieldSmallButtonContainer.cs" />
+    <Compile Include="Controls\BasePathEditorControl.cs" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Controls\" />
